backend/requirements.txt
# Framework Web
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==23.0.0

# Base de datos
sqlalchemy==2.0.35
psycopg[binary]==3.2.6

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Validación
pydantic==2.9.0
pydantic-settings==2.5.0

# Utilidades
python-dotenv==1.0.1
httpx==0.27.0
python-dateutil==2.9.0

# Cliente Supabase (OPCIONAL - instalar después si lo necesitas)
# supabase==2.9.0

# IA - Claude API (OPCIONAL - instalar después si lo necesitas)
# anthropic==0.39.0